requirements.txt
alembic==1.4.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.2
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-User==0.6.21
Flask-WTF==0.14.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
passlib==1.7.2
pbr==5.4.4
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.21.0
six==1.14.0
SQLAlchemy==1.3.13
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
urllib3==1.24.3
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
